backend/requirements.txt
fastapi
uvicorn
python-dotenv
pinecone
sentence-transformers
pandas
tqdm
langchain
kaggle
kagglehub
langchain-community
langchain-huggingface